Move misplaced functions to their respective _example.go files
