+++ internal/sdkv2/morpheus/resources/integration/morpheus_integration_ansible_tower_resource_example.go
@@ -2,4 +2,50 @@
 
 package integration
 
+import (
+	"fmt"
+	"path/filepath"
+	"runtime"
+	"testing"
+
+	"github.com/HPE/terraform-provider-hpe/internal/framework/subproviders/morpheus/testhelpers"
+)
+
 //go:generate go run ../../../../../cmd/render -out examples/resources/morpheus_integration_ansible_tower/resource.tf morpheus_integration_ansible_tower_resource.tf.tmpl Name "tfexample ansible tower integration" Enabled true Url "https://ansibletower01.morpheusdata.com" Username admin Password password123
+
+// RenderMorpheusIntegrationAnsibleTowerConfig renders the Ansible Tower integration resource configuration
+// with default values that can be overridden via the overrides parameter.
+func RenderMorpheusIntegrationAnsibleTowerConfig(t *testing.T, overrides map[string]string) (string, error) {
+	t.Helper()
+
+	defaults := map[string]string{
+		"Name":     "tf_test_ansible_tower",
+		"Enabled":  "true",
+		"Url":      "https://ansibletower01.morpheusdata.com",
+		"Username": "admin",
+		"Password": "password123",
+	}
+
+	// Override defaults with provided values
+	for key, value := range overrides {
+		defaults[key] = value
+	}
+
+	// Get the directory where this source file is located
+	_, filename, _, ok := runtime.Caller(0)
+	if !ok {
+		return "", fmt.Errorf("unable to get current file path")
+	}
+	dir := filepath.Dir(filename)
+	templatePath := filepath.Join(dir, "morpheus_integration_ansible_tower_resource.tf.tmpl")
+
+	return testhelpers.RenderExample(
+		t,
+		templatePath,
+		"Name", defaults["Name"],
+		"Enabled", defaults["Enabled"],
+		"Url", defaults["Url"],
+		"Username", defaults["Username"],
+		"Password", defaults["Password"],
+	)
+}
