+++ internal/sdkv2/morpheus/resources/optiontype/morpheus_option_type_password_resource_example.go
@@ -6,68 +6,13 @@ import (
 	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
 
 	"fmt"
+	"github.com/HPE/terraform-provider-hpe/internal/framework/subproviders/morpheus/testhelpers"
 	"path/filepath"
 	"runtime"
 	"testing"
-	"github.com/HPE/terraform-provider-hpe/internal/framework/subproviders/morpheus/testhelpers"
 )
 
 //go:generate go run ../../../../../cmd/render -out examples/resources/morpheus_option_type_password/resource.tf morpheus_option_type_password_resource.tf.tmpl Name tf_example_password_option_type Description "Terraform password option type example" Labels "[\"demo\", \"terraform\"]" FieldName tfPasswordExample ExportMeta true DependentField dependent_example VisibilityField visibility_example RequireField require_example ShowOnEdit true Editable true DisplayValueOnDetails true FieldLabel numbers Placeholder fewf DefaultValue testing HelpBlock fiwefw Required true VerifyPattern "a\\\\D{4}"
-
-func RenderMorpheusOptionTypeTextareaConfig(t *testing.T, overrides map[string]string) string {
-	t.Helper()
-
-	defaults := map[string]string{
-		"Name":                  acctest.RandomWithPrefix(t.Name()),
-		"Description":           "Terraform text area option type example",
-		"Labels":                `["demo","terraform"]`,
-		"FieldName":             "textareaExample",
-		"ExportMeta":            "true",
-		"DependentField":        "dependent_example",
-		"VisibilityField":       "visibility_example",
-		"RequireField":          "require_example",
-		"ShowOnEdit":            "true",
-		"Editable":              "true",
-		"DisplayValueOnDetails": "true",
-		"FieldLabel":            "Text Area Example",
-		"Rows":                  "5",
-		"Placeholder":           "example text",
-		"DefaultValue":          "example",
-		"HelpBlock":             "Terraform text area option type example",
-		"Required":              "true",
-		"VerifyPattern":         `a\\D{4}`,
-	}
-
-	for key, value := range overrides {
-		defaults[key] = value
-	}
-
-	resourceConfig, err := testhelpers.RenderExample(t, "morpheus_option_type_textarea_resource.tf.tmpl",
-		"Name", defaults["Name"],
-		"Description", defaults["Description"],
-		"Labels", defaults["Labels"],
-		"FieldName", defaults["FieldName"],
-		"ExportMeta", defaults["ExportMeta"],
-		"DependentField", defaults["DependentField"],
-		"VisibilityField", defaults["VisibilityField"],
-		"RequireField", defaults["RequireField"],
-		"ShowOnEdit", defaults["ShowOnEdit"],
-		"Editable", defaults["Editable"],
-		"DisplayValueOnDetails", defaults["DisplayValueOnDetails"],
-		"FieldLabel", defaults["FieldLabel"],
-		"Rows", defaults["Rows"],
-		"Placeholder", defaults["Placeholder"],
-		"DefaultValue", defaults["DefaultValue"],
-		"HelpBlock", defaults["HelpBlock"],
-		"Required", defaults["Required"],
-		"VerifyPattern", defaults["VerifyPattern"],
-	)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	return resourceConfig
-}
 
 // RenderMorpheusOptionTypePasswordConfig generates a Terraform configuration for the
 // morpheus_option_type_password resource.
@@ -133,176 +78,3 @@ func RenderMorpheusOptionTypePasswordConfig(t *testing.T, overrides map[string]s
 		"VerifyPattern", defaults["VerifyPattern"],
 	)
 }
-
-// RenderMorpheusOptionTypeNumberConfig generates a Terraform configuration for the
-// morpheus_option_type_number resource. It accepts an optional map of field overrides to
-// customize the default values.
-// Supported override keys: "Name", "Description", "Labels", "FieldName", "ExportMeta",
-// "DependentField", "VisibilityField", "RequireField", "ShowOnEdit", "Editable",
-// "DisplayValueOnDetails", "FieldLabel", "Placeholder", "DefaultValue", "HelpBlock", "Required"
-func RenderMorpheusOptionTypeNumberConfig(t *testing.T, overrides map[string]string) (string, error) {
-	t.Helper()
-
-	defaults := map[string]string{
-		"Name":                  acctest.RandomWithPrefix(t.Name()),
-		"Description":           "Terraform number option type example",
-		"Labels":                "[\"demo\", \"terraform\"]",
-		"FieldName":             "tfNumberExample",
-		"ExportMeta":            "true",
-		"DependentField":        "dependent_example",
-		"VisibilityField":       "visibility_example",
-		"RequireField":          "require_example",
-		"ShowOnEdit":            "true",
-		"Editable":              "true",
-		"DisplayValueOnDetails": "true",
-		"FieldLabel":            "Number Example",
-		"Placeholder":           "12",
-		"DefaultValue":          "1",
-		"HelpBlock":             "Provide a number",
-		"Required":              "true",
-	}
-
-	// Apply overrides
-	for key, value := range overrides {
-		defaults[key] = value
-	}
-
-	// Get the directory where this source file is located
-	_, filename, _, ok := runtime.Caller(0)
-	if !ok {
-		return "", fmt.Errorf("unable to get current file path")
-	}
-	dir := filepath.Dir(filename)
-	templatePath := filepath.Join(dir, "morpheus_option_type_number_resource.tf.tmpl")
-
-	return testhelpers.RenderExample(
-		t,
-		templatePath,
-		"Name", defaults["Name"],
-		"Description", defaults["Description"],
-		"Labels", defaults["Labels"],
-		"FieldName", defaults["FieldName"],
-		"ExportMeta", defaults["ExportMeta"],
-		"DependentField", defaults["DependentField"],
-		"VisibilityField", defaults["VisibilityField"],
-		"RequireField", defaults["RequireField"],
-		"ShowOnEdit", defaults["ShowOnEdit"],
-		"Editable", defaults["Editable"],
-		"DisplayValueOnDetails", defaults["DisplayValueOnDetails"],
-		"FieldLabel", defaults["FieldLabel"],
-		"Placeholder", defaults["Placeholder"],
-		"DefaultValue", defaults["DefaultValue"],
-		"HelpBlock", defaults["HelpBlock"],
-		"Required", defaults["Required"],
-	)
-}
-
-// RenderMorpheusOptionTypeHiddenConfig generates a Terraform configuration for the
-// morpheus_option_type_hidden resource. It accepts an optional map of field overrides to customize
-// the default values. Supported override keys: "Name", "Description", "Labels", "FieldName",
-// "ExportMeta", "DependentField", "VisibilityField", "RequireField", "ShowOnEdit", "Editable",
-// "DisplayValueOnDetails", "DefaultValue"
-func RenderMorpheusOptionTypeHiddenConfig(t *testing.T, overrides map[string]string) (string, error) {
-	t.Helper()
-
-	defaults := map[string]string{
-		"Name":                  acctest.RandomWithPrefix(t.Name()),
-		"Description":           "Terraform hidden option type example",
-		"Labels":                `["demo","terraform"]`,
-		"FieldName":             "hidden_example",
-		"ExportMeta":            "true",
-		"DependentField":        "dependent_example",
-		"VisibilityField":       "visibility_example",
-		"RequireField":          "require_example",
-		"ShowOnEdit":            "true",
-		"Editable":              "true",
-		"DisplayValueOnDetails": "true",
-		"DefaultValue":          "example",
-	}
-
-	// Apply overrides
-	for key, value := range overrides {
-		defaults[key] = value
-	}
-
-	// Get the directory where this source file is located
-	_, filename, _, ok := runtime.Caller(0)
-	if !ok {
-		return "", fmt.Errorf("unable to get current file path")
-	}
-	dir := filepath.Dir(filename)
-	templatePath := filepath.Join(dir, "morpheus_option_type_hidden_resource.tf.tmpl")
-
-	return testhelpers.RenderExample(
-		t,
-		templatePath,
-		"Name", defaults["Name"],
-		"Description", defaults["Description"],
-		"Labels", defaults["Labels"],
-		"FieldName", defaults["FieldName"],
-		"ExportMeta", defaults["ExportMeta"],
-		"DependentField", defaults["DependentField"],
-		"VisibilityField", defaults["VisibilityField"],
-		"RequireField", defaults["RequireField"],
-		"ShowOnEdit", defaults["ShowOnEdit"],
-		"Editable", defaults["Editable"],
-		"DisplayValueOnDetails", defaults["DisplayValueOnDetails"],
-		"DefaultValue", defaults["DefaultValue"],
-	)
-}
-
-// RenderOptionTypeCheckboxConfig generates a Terraform configuration for the morpheus_option_type_checkbox resource.
-// It accepts an optional map of field overrides to customize the default values.
-// Supported override keys: "Name", "Description", "Labels", "FieldName", "ExportMeta", "DependentField",
-// "VisibilityField", "RequireField", "ShowOnEdit", "Editable", "DisplayValueOnDetails", "FieldLabel", "DefaultChecked"
-func RenderOptionTypeCheckboxConfig(t *testing.T, overrides map[string]string) (string, error) {
-	t.Helper()
-
-	defaults := map[string]string{
-		"Name":                  acctest.RandomWithPrefix(t.Name()),
-		"Description":           "Terraform checkbox option type example",
-		"Labels":                "[\"demo\", \"terraform\"]",
-		"FieldName":             "checkbox_example",
-		"ExportMeta":            "true",
-		"DependentField":        "dependent_example",
-		"VisibilityField":       "visibility_example",
-		"RequireField":          "require_example",
-		"ShowOnEdit":            "true",
-		"Editable":              "true",
-		"DisplayValueOnDetails": "true",
-		"FieldLabel":            "Checkbox Example",
-		"DefaultChecked":        "true",
-	}
-
-	// Apply overrides
-	for key, value := range overrides {
-		defaults[key] = value
-	}
-
-	// Get the directory where this source file is located
-	_, filename, _, ok := runtime.Caller(0)
-	if !ok {
-		return "", fmt.Errorf("unable to get current file path")
-	}
-	dir := filepath.Dir(filename)
-	templatePath := filepath.Join(dir, "morpheus_option_type_checkbox_resource.tf.tmpl")
-
-	return testhelpers.RenderExample(
-		t,
-		templatePath,
-		"Name", defaults["Name"],
-		"Description", defaults["Description"],
-		"Labels", defaults["Labels"],
-		"FieldName", defaults["FieldName"],
-		"ExportMeta", defaults["ExportMeta"],
-		"DependentField", defaults["DependentField"],
-		"VisibilityField", defaults["VisibilityField"],
-		"RequireField", defaults["RequireField"],
-		"ShowOnEdit", defaults["ShowOnEdit"],
-		"Editable", defaults["Editable"],
-		"DisplayValueOnDetails", defaults["DisplayValueOnDetails"],
-		"FieldLabel", defaults["FieldLabel"],
-		"DefaultChecked", defaults["DefaultChecked"],
-	)
-}
-
