+++ internal/sdkv2/morpheus/resources/task/morpheus_task_groovy_script_resource_example.go
@@ -2,6 +2,161 @@
 
 package task
 
+import (
+	"fmt"
+	"path/filepath"
+	"runtime"
+	"testing"
+
+	"github.com/HPE/terraform-provider-hpe/internal/framework/subproviders/morpheus/testhelpers"
+)
+
 //go:generate sh -c "go run ../../../../../cmd/render -out examples/resources/morpheus_task_groovy_script/resource.tf morpheus_task_groovy_script_resource.tf.tmpl Name tfexample_groovy_local Code tfexample_groovy_local Labels '[\"demo\",\"terraform\"]' SourceType local ScriptContent 'println \"hello\"' Retryable true RetryCount 1 RetryDelaySeconds 10 AllowCustomConfig true"
 //go:generate sh -c "go run ../../../../../cmd/render -out examples/resources/morpheus_task_groovy_script/resource_git.tf morpheus_task_groovy_script_resource_git.tf.tmpl Name tfexample_groovy_git Code tfexample_groovy_git Labels '[\"demo\",\"terraform\"]' SourceType repository ResultType json ScriptPath example.groovy VersionRef master RepositoryId 1 Retryable true RetryCount 1 RetryDelaySeconds 10 AllowCustomConfig true"
 //go:generate sh -c "go run ../../../../../cmd/render -out examples/resources/morpheus_task_groovy_script/resource_url.tf morpheus_task_groovy_script_resource_url.tf.tmpl Name tfexample_groovy_url Code tfexample_groovy_url Labels '[\"demo\",\"terraform\"]' SourceType url ResultType json ScriptPath https://example.com/example.groovy Retryable true RetryCount 1 RetryDelaySeconds 10 AllowCustomConfig true"
+
+func RenderMorpheusTaskGroovyScriptConfig(
+	t *testing.T,
+	name string,
+	overrides map[string]string,
+) (string, error) {
+	t.Helper()
+
+	defaults := map[string]string{
+		"Name":              name,
+		"Code":              name,
+		"Labels":            "[\"demo\", \"terraform\"]",
+		"SourceType":        "local",
+		"ScriptContent":     "println \"hello\"",
+		"Retryable":         "true",
+		"RetryCount":        "1",
+		"RetryDelaySeconds": "10",
+		"AllowCustomConfig": "true",
+	}
+
+	for key, value := range overrides {
+		defaults[key] = value
+	}
+
+	// Get the directory where this source file is located
+	_, filename, _, ok := runtime.Caller(0)
+	if !ok {
+		return "", fmt.Errorf("unable to get current file path")
+	}
+	dir := filepath.Dir(filename)
+	templatePath := filepath.Join(dir, "morpheus_task_groovy_script_resource.tf.tmpl")
+
+	return testhelpers.RenderExample(
+		t,
+		templatePath,
+		"Name", defaults["Name"],
+		"Code", defaults["Code"],
+		"Labels", defaults["Labels"],
+		"SourceType", defaults["SourceType"],
+		"ScriptContent", defaults["ScriptContent"],
+		"Retryable", defaults["Retryable"],
+		"RetryCount", defaults["RetryCount"],
+		"RetryDelaySeconds", defaults["RetryDelaySeconds"],
+		"AllowCustomConfig", defaults["AllowCustomConfig"],
+	)
+}
+
+func RenderMorpheusTaskGroovyScriptGitConfig(
+	t *testing.T,
+	name string,
+	overrides map[string]string,
+) (string, error) {
+	t.Helper()
+
+	defaults := map[string]string{
+		"Name":              name,
+		"Code":              name,
+		"Labels":            "[\"demo\", \"terraform\"]",
+		"SourceType":        "repository",
+		"ResultType":        "json",
+		"ScriptPath":        "example.groovy",
+		"VersionRef":        "master",
+		"RepositoryId":      "1",
+		"Retryable":         "true",
+		"RetryCount":        "1",
+		"RetryDelaySeconds": "10",
+		"AllowCustomConfig": "true",
+	}
+
+	for key, value := range overrides {
+		defaults[key] = value
+	}
+
+	// Get the directory where this source file is located
+	_, filename, _, ok := runtime.Caller(0)
+	if !ok {
+		return "", fmt.Errorf("unable to get current file path")
+	}
+	dir := filepath.Dir(filename)
+	templatePath := filepath.Join(dir, "morpheus_task_groovy_script_resource_git.tf.tmpl")
+
+	return testhelpers.RenderExample(
+		t,
+		templatePath,
+		"Name", defaults["Name"],
+		"Code", defaults["Code"],
+		"Labels", defaults["Labels"],
+		"SourceType", defaults["SourceType"],
+		"ResultType", defaults["ResultType"],
+		"ScriptPath", defaults["ScriptPath"],
+		"VersionRef", defaults["VersionRef"],
+		"RepositoryId", defaults["RepositoryId"],
+		"Retryable", defaults["Retryable"],
+		"RetryCount", defaults["RetryCount"],
+		"RetryDelaySeconds", defaults["RetryDelaySeconds"],
+		"AllowCustomConfig", defaults["AllowCustomConfig"],
+	)
+}
+
+func RenderMorpheusTaskGroovyScriptUrlConfig(
+	t *testing.T,
+	name string,
+	overrides map[string]string,
+) (string, error) {
+	t.Helper()
+
+	defaults := map[string]string{
+		"Name":              name,
+		"Code":              name,
+		"Labels":            "[\"demo\", \"terraform\"]",
+		"SourceType":        "url",
+		"ResultType":        "json",
+		"ScriptPath":        "https://example.com/example.groovy",
+		"Retryable":         "true",
+		"RetryCount":        "1",
+		"RetryDelaySeconds": "10",
+		"AllowCustomConfig": "true",
+	}
+
+	for key, value := range overrides {
+		defaults[key] = value
+	}
+
+	// Get the directory where this source file is located
+	_, filename, _, ok := runtime.Caller(0)
+	if !ok {
+		return "", fmt.Errorf("unable to get current file path")
+	}
+	dir := filepath.Dir(filename)
+	templatePath := filepath.Join(dir, "morpheus_task_groovy_script_resource_url.tf.tmpl")
+
+	return testhelpers.RenderExample(
+		t,
+		templatePath,
+		"Name", defaults["Name"],
+		"Code", defaults["Code"],
+		"Labels", defaults["Labels"],
+		"SourceType", defaults["SourceType"],
+		"ResultType", defaults["ResultType"],
+		"ScriptPath", defaults["ScriptPath"],
+		"Retryable", defaults["Retryable"],
+		"RetryCount", defaults["RetryCount"],
+		"RetryDelaySeconds", defaults["RetryDelaySeconds"],
+		"AllowCustomConfig", defaults["AllowCustomConfig"],
+	)
+}
