+++ internal/sdkv2/morpheus/resources/task/morpheus_task_javascript_resource_example.go
@@ -2,4 +2,60 @@
 
 package task
 
+import (
+	"fmt"
+	"path/filepath"
+	"runtime"
+	"testing"
+
+	"github.com/HPE/terraform-provider-hpe/internal/framework/subproviders/morpheus/testhelpers"
+)
+
 //go:generate go run ../../../../../cmd/render -out examples/resources/morpheus_task_javascript/resource.tf morpheus_task_javascript_resource.tf.tmpl Name tfexample_javascript Code tfexample_javascript Labels ["demo","terraform"] ScriptContent console.log("testing") Retryable true RetryCount 1 RetryDelaySeconds 10 AllowCustomConfig true
+
+// RenderTaskJavascriptConfig generates a terraform configuration string
+// for task javascript resource.
+// It accepts a name and a map to override default field values.
+func RenderTaskJavascriptConfig(
+	t *testing.T,
+	name string,
+	overrides map[string]string,
+) (string, error) {
+	t.Helper()
+
+	defaults := map[string]string{
+		"Name":              name,
+		"Code":              name,
+		"Labels":            `["demo","terraform"]`,
+		"ScriptContent":     `console.log("testing")`,
+		"Retryable":         "true",
+		"RetryCount":        "1",
+		"RetryDelaySeconds": "10",
+		"AllowCustomConfig": "true",
+	}
+
+	for key, value := range overrides {
+		defaults[key] = value
+	}
+
+	// Get the directory where this source file is located
+	_, filename, _, ok := runtime.Caller(0)
+	if !ok {
+		return "", fmt.Errorf("unable to get current file path")
+	}
+	dir := filepath.Dir(filename)
+	templatePath := filepath.Join(dir, "morpheus_task_javascript_resource.tf.tmpl")
+
+	return testhelpers.RenderExample(
+		t,
+		templatePath,
+		"Name", defaults["Name"],
+		"Code", defaults["Code"],
+		"Labels", defaults["Labels"],
+		"ScriptContent", defaults["ScriptContent"],
+		"Retryable", defaults["Retryable"],
+		"RetryCount", defaults["RetryCount"],
+		"RetryDelaySeconds", defaults["RetryDelaySeconds"],
+		"AllowCustomConfig", defaults["AllowCustomConfig"],
+	)
+}
