+++ internal/sdkv2/morpheus/resources/task/morpheus_task_restart_resource_example.go
@@ -2,4 +2,59 @@
 
 package task
 
+import (
+	"fmt"
+	"path/filepath"
+	"runtime"
+	"testing"
+
+	"github.com/HPE/terraform-provider-hpe/internal/framework/subproviders/morpheus/testhelpers"
+)
+
 //go:generate go run ../../../../../cmd/render -out examples/resources/morpheus_task_restart/resource.tf morpheus_task_restart_resource.tf.tmpl Name tfexample_restart Code tfexample_restart Labels "[\"demo\", \"terraform\"]" Retryable true RetryCount 1 RetryDelaySeconds 10 AllowCustomConfig true
+
+// RenderTaskRestartConfig renders the task restart resource configuration
+// with the provided name and field overrides.
+func RenderTaskRestartConfig(
+	t *testing.T,
+	name string,
+	overrides map[string]string,
+) (string, error) {
+	t.Helper()
+
+	// Default values
+	defaults := map[string]string{
+		"Name":              name,
+		"Code":              "tfexample_restart",
+		"Labels":            `["demo", "terraform"]`,
+		"Retryable":         "true",
+		"RetryCount":        "1",
+		"RetryDelaySeconds": "10",
+		"AllowCustomConfig": "true",
+	}
+
+	// Apply overrides
+	for key, value := range overrides {
+		defaults[key] = value
+	}
+
+	// Get the directory where this source file is located
+	_, filename, _, ok := runtime.Caller(0)
+	if !ok {
+		return "", fmt.Errorf("unable to get current file path")
+	}
+	dir := filepath.Dir(filename)
+	templatePath := filepath.Join(dir, "morpheus_task_restart_resource.tf.tmpl")
+
+	return testhelpers.RenderExample(
+		t,
+		templatePath,
+		"Name", defaults["Name"],
+		"Code", defaults["Code"],
+		"Labels", defaults["Labels"],
+		"Retryable", defaults["Retryable"],
+		"RetryCount", defaults["RetryCount"],
+		"RetryDelaySeconds", defaults["RetryDelaySeconds"],
+		"AllowCustomConfig", defaults["AllowCustomConfig"],
+	)
+}
