+++ internal/sdkv2/morpheus/resources/task/morpheus_task_shell_script_example.go
@@ -2,6 +2,167 @@
 
 package task
 
+import (
+	"fmt"
+	"path/filepath"
+	"runtime"
+	"testing"
+
+	"github.com/HPE/terraform-provider-hpe/internal/framework/subproviders/morpheus/testhelpers"
+)
+
 //go:generate sh -c "cd ../../../../../internal/sdkv2/morpheus/resources/task && go run ../../../../../cmd/render/main.go -out examples/resources/morpheus_task_shell_script/resource.tf morpheus_task_shell_script_resource.tf.tmpl Name tfexample_shell_local Code tfexample_shell_local Labels '[\"demo\", \"terraform\"]' SourceType local ScriptContent '  echo \"testing\"' Sudo true Retryable true RetryCount 1 RetryDelaySeconds 10 AllowCustomConfig true"
 //go:generate sh -c "cd ../../../../../internal/sdkv2/morpheus/resources/task && go run ../../../../../cmd/render/main.go -out examples/resources/morpheus_task_shell_script/resource_git.tf morpheus_task_shell_script_resource_git.tf.tmpl Name tfexample_shell_git Code tfexample_shell_git Labels '[\"demo\", \"terraform\"]' SourceType repository ResultType json ScriptPath example.sh VersionRef master RepositoryId 1 Sudo true Retryable true RetryCount 1 RetryDelaySeconds 10 AllowCustomConfig true"
 //go:generate sh -c "cd ../../../../../internal/sdkv2/morpheus/resources/task && go run ../../../../../cmd/render/main.go -out examples/resources/morpheus_task_shell_script/resource_url.tf morpheus_task_shell_script_resource_url.tf.tmpl Name tfexample_shell_url Code tfexample_shell_url Labels '[\"demo\", \"terraform\"]' SourceType url ResultType json ScriptPath https://example.com/example.sh Sudo true Retryable true RetryCount 1 RetryDelaySeconds 10 AllowCustomConfig true"
+
+func RenderTaskShellScriptConfig(
+	t *testing.T,
+	name string,
+	overrides map[string]string,
+) (string, error) {
+	t.Helper()
+
+	defaults := map[string]string{
+		"Name":              name,
+		"Code":              name,
+		"Labels":            "[\"demo\", \"terraform\"]",
+		"SourceType":        "local",
+		"ScriptContent":     "  echo \"testing\"",
+		"Sudo":              "true",
+		"Retryable":         "true",
+		"RetryCount":        "1",
+		"RetryDelaySeconds": "10",
+		"AllowCustomConfig": "true",
+	}
+
+	for key, value := range overrides {
+		defaults[key] = value
+	}
+
+	// Get the directory where this source file is located
+	_, filename, _, ok := runtime.Caller(0)
+	if !ok {
+		return "", fmt.Errorf("unable to get current file path")
+	}
+	dir := filepath.Dir(filename)
+	templatePath := filepath.Join(dir, "morpheus_task_shell_script_resource.tf.tmpl")
+
+	return testhelpers.RenderExample(
+		t,
+		templatePath,
+		"Name", defaults["Name"],
+		"Code", defaults["Code"],
+		"Labels", defaults["Labels"],
+		"SourceType", defaults["SourceType"],
+		"ScriptContent", defaults["ScriptContent"],
+		"Sudo", defaults["Sudo"],
+		"Retryable", defaults["Retryable"],
+		"RetryCount", defaults["RetryCount"],
+		"RetryDelaySeconds", defaults["RetryDelaySeconds"],
+		"AllowCustomConfig", defaults["AllowCustomConfig"],
+	)
+}
+
+func RenderTaskShellScriptGitConfig(
+	t *testing.T,
+	name string,
+	overrides map[string]string,
+) (string, error) {
+	t.Helper()
+
+	defaults := map[string]string{
+		"Name":              name,
+		"Code":              name,
+		"Labels":            "[\"demo\", \"terraform\"]",
+		"SourceType":        "repository",
+		"ResultType":        "json",
+		"ScriptPath":        "example.sh",
+		"VersionRef":        "master",
+		"RepositoryId":      "0",
+		"Sudo":              "true",
+		"Retryable":         "true",
+		"RetryCount":        "1",
+		"RetryDelaySeconds": "10",
+		"AllowCustomConfig": "true",
+	}
+
+	for key, value := range overrides {
+		defaults[key] = value
+	}
+
+	// Get the directory where this source file is located
+	_, filename, _, ok := runtime.Caller(0)
+	if !ok {
+		return "", fmt.Errorf("unable to get current file path")
+	}
+	dir := filepath.Dir(filename)
+	templatePath := filepath.Join(dir, "morpheus_task_shell_script_resource_git.tf.tmpl")
+
+	return testhelpers.RenderExample(
+		t,
+		templatePath,
+		"Name", defaults["Name"],
+		"Code", defaults["Code"],
+		"Labels", defaults["Labels"],
+		"SourceType", defaults["SourceType"],
+		"ResultType", defaults["ResultType"],
+		"ScriptPath", defaults["ScriptPath"],
+		"VersionRef", defaults["VersionRef"],
+		"RepositoryId", defaults["RepositoryId"],
+		"Sudo", defaults["Sudo"],
+		"Retryable", defaults["Retryable"],
+		"RetryCount", defaults["RetryCount"],
+		"RetryDelaySeconds", defaults["RetryDelaySeconds"],
+		"AllowCustomConfig", defaults["AllowCustomConfig"],
+	)
+}
+
+func RenderTaskShellScriptUrlConfig(
+	t *testing.T,
+	name string,
+	overrides map[string]string,
+) (string, error) {
+	t.Helper()
+
+	defaults := map[string]string{
+		"Name":              name,
+		"Code":              name,
+		"Labels":            "[\"demo\", \"terraform\"]",
+		"SourceType":        "url",
+		"ResultType":        "json",
+		"ScriptPath":        "https://example.com/example.sh",
+		"Sudo":              "true",
+		"Retryable":         "true",
+		"RetryCount":        "1",
+		"RetryDelaySeconds": "10",
+		"AllowCustomConfig": "true",
+	}
+
+	for key, value := range overrides {
+		defaults[key] = value
+	}
+
+	// Get the directory where this source file is located
+	_, filename, _, ok := runtime.Caller(0)
+	if !ok {
+		return "", fmt.Errorf("unable to get current file path")
+	}
+	dir := filepath.Dir(filename)
+	templatePath := filepath.Join(dir, "morpheus_task_shell_script_resource_url.tf.tmpl")
+
+	return testhelpers.RenderExample(
+		t,
+		templatePath,
+		"Name", defaults["Name"],
+		"Code", defaults["Code"],
+		"Labels", defaults["Labels"],
+		"SourceType", defaults["SourceType"],
+		"ResultType", defaults["ResultType"],
+		"ScriptPath", defaults["ScriptPath"],
+		"Sudo", defaults["Sudo"],
+		"Retryable", defaults["Retryable"],
+		"RetryCount", defaults["RetryCount"],
+		"RetryDelaySeconds", defaults["RetryDelaySeconds"],
+		"AllowCustomConfig", defaults["AllowCustomConfig"],
+	)
+}
