+++ internal/sdkv2/morpheus/resources/template/morpheus_script_template_example.go
@@ -2,4 +2,56 @@
 
 package template
 
+import (
+	"fmt"
+	"path/filepath"
+	"runtime"
+	"testing"
+
+	"github.com/HPE/terraform-provider-hpe/internal/framework/subproviders/morpheus/testhelpers"
+)
+
 //go:generate go run ../../../../../cmd/render -out examples/resources/morpheus_script_template/resource.tf morpheus_script_template_resource.tf.tmpl 'Name' 'tf-terraform-script-template' 'Labels' "[\"demo\", \"template\", \"terraform\"]" 'ScriptType' 'bash' 'ScriptPhase' 'provision' 'ScriptContent' "echo \"testing\"" 'RunAsUser' 'root' 'Sudo' 'true'
+
+// RenderScriptTemplateConfig renders the template with provided overrides
+func RenderScriptTemplateConfig(
+	t *testing.T,
+	name string,
+	overrides map[string]string,
+) (string, error) {
+	t.Helper()
+
+	defaults := map[string]string{
+		"Name":          name,
+		"Labels":        "[\"demo\", \"template\", \"terraform\"]",
+		"ScriptType":    "bash",
+		"ScriptPhase":   "provision",
+		"ScriptContent": "echo \"testing\"",
+		"RunAsUser":     "root",
+		"Sudo":          "true",
+	}
+
+	for key, value := range overrides {
+		defaults[key] = value
+	}
+
+	// Get the directory where this source file is located
+	_, filename, _, ok := runtime.Caller(0)
+	if !ok {
+		return "", fmt.Errorf("unable to get current file path")
+	}
+	dir := filepath.Dir(filename)
+	templatePath := filepath.Join(dir, "morpheus_script_template_resource.tf.tmpl")
+
+	return testhelpers.RenderExample(
+		t,
+		templatePath,
+		"Name", defaults["Name"],
+		"Labels", defaults["Labels"],
+		"ScriptType", defaults["ScriptType"],
+		"ScriptPhase", defaults["ScriptPhase"],
+		"ScriptContent", defaults["ScriptContent"],
+		"RunAsUser", defaults["RunAsUser"],
+		"Sudo", defaults["Sudo"],
+	)
+}
