+++ internal/sdkv2/morpheus/resources/trust/morpheus_key_pair_resource_example.go
@@ -65,4 +65,3 @@ func RenderMorpheusKeyPairConfig(t *testing.T, overrides map[string]string) (str
 		"PrivateKey", defaults["PrivateKey"],
 	)
 }
-
